On Linux, /proc/filesystems lists the currently available filesystems that the OS supports. This list may change dynamically as kernel modules are added/removed.
On other platforms, there may not be a /proc/filesystems file; the program GetSupportedFilesystems works around this by implementing platform-specific equivalents. For each supported filesystem, it outputs the vfstype (that would be used as argument to the 'mount -t' command) and a newline character to standard output.

Unless a specific implementation exists, GetSupportedFilesystems falls back to the default, which is to parse /proc/filesystems. If /proc/filesystems is missing, nothing will be output. From the Script package point of view, such an empty list of supported filesystems will result in a FiboSandbox when installing recipes, which should be acceptable.

Patch by David Karell.