Add Python module for Aliens, query-only at this stage.
This module can answer several questions about the state of alien
packages in the system, and also return information to CheckDependencies
about the alien system itself.

The intention is for this to be fleshed out into a script later, which
can then be called from Compile, and to add an --install option to
trigger installation of an alien package.