Add Alien querying support to CheckDependencies.
This replaces the proof-of-concept code accidentally committed in r4073
with safe versions using the Alien module added in the last commit.

With this change, CheckDependencies will process dependency rules of the
form <AlienManager>:<AlienPackage> by handing off to the Alien module to
query the <AlienManager> wrapper for details about <AlienPackage>, and
add the alien match to the matches list if the dependency is not already
met. If the alien manager is not installed at all, it will also add that
to the matches before any alien packages that require it.