Add Alien-LuaRocks script, implementing the Aliens interface.
This script gives an example of the proposed interface and has been
written for clarity in that mor than for any other goal. It is likely
that the final interface script will be in Lua and access the LuaRocks
database directly, rather than parsing the output of the command-line
tool. The same interface is proposed for every wrapper of an alien
package manager.

This script does not support installing yet, and does not specifically
use the /System/Aliens tree nor create the /S/A/LuaRocks hierarchy.