Some packages like KDE and CUPS are composed of multiple software that are compiled and installed on different prefixes. It's quite hard to keep modifying these software's build variables to guarantee that it can see libs from another software, though, especially when these software's build system is complex. (think of cmake)
We have solved a similar problem before with the introduction of $target/Shared and $target/share -> /System/Links/Shared. That's very intrusive, though, requiring changes to binary packages, Scripts and Compile.

For compiled packages, LD_PRELOAD seems to work quite good, at least for the KDE and CUPS environments. At some point we should have a wrapper to cupsd and to startkde executables so that they preload this library. Comments are welcome.