Symlink bin/ entries from Aliens.
This should really be in the main Alien script and
not in the language-specific drivers, but oh well.
The main Alien script should probably be broken into
a Alien.py module for CheckDependencies and a thin
launcher (since all that __main__ does is to call
the appropriate driver).

Also, let pip install over an existing version.

Also, beginnings of a remove function for convenience.