Add initial version of Alien-CTAN.
CTAN doesn't have a command-line client, just a repository website,
so this Alien driver scrapes the website and performs some
minimal package management on its own.

It currently installs to the default TeX path, not to /System/Aliens.
In my system, it is defaulting to $HOME/texmf. Tested with the
ancient TeTeX 3.0, which is the TeX distro that I use to this day.

This should be fixable in a future TeX distribution for GoboLinux,
configuring TeX to add a new path like /System/Aliens/CTAN as its
lookup path for files. Then we could configure Alien-CTAN to use it.