Runner: include Bash in the overlay if needed.
When Runner is invoked with '--pure', a complete list of all programs
that will be exposed under /System/Index needs to be provided in the
Dependencies files. However, under certain circumstances, Runner
itself may end up adding a dependency on Bash as part of the creation
of a wrapper script; such script may be needed to source special files
(like Resources/Environment) prior to running the executable.

This commit ensures that Bash and its direct dependencies are included
in the overlay on such cases.