Let $scriptPath deal with overlayfs.
When a program is overlay-mounted over /System/Index that program's
executables are not represented as symlinks pointing to the
corresponding entries on /Programs. Rather, they're shown as regular
files, so $scriptPath will not work as designed.

This patch makes $scriptPath deal with overlayfs so that program
name lookups work as expected when Runner is into play. With this
change we no longer need to keep an ignore-list in Runner, either,
so this commit modifies that as well.