PostInstall: Partially revert 06122c1
Do not delete the whole $goboData directory, as PostInstall potentially installs files into that directory! Previously we've had 'rmdir --ignore-fail-on-non-empty ./run' in there which seemed defunct, so not adding it back

I did *previously* run into issues with leftover files at /Data/Variable/run/mount, but cannot reproduce anymore. It was *possibly* related to what was addressed in 4dffc45, but let's keep an eye on it...